requirements.txt
paho-mqtt
influxdb3-python
python-dotenv